requirements.txt
# Discord Bot
discord.py>=2.3.0

# Stock Data
yfinance>=0.2.40

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
mplfinance>=0.12.10b0

# LLM for Sentiment Analysis
openai>=1.0.0

# ML Models (Advanced)
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow>=2.15.0  # Optional - for LSTM

# Scheduling
schedule>=1.2.0

# Environment
python-dotenv>=1.0.0

# Phase 0 — persistence & integrations
supabase>=2.0.0
httpx>=0.25.0
aiohttp>=3.9.0
tenacity>=8.2.0
pandas-market-calendars>=4.3.0
structlog>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
anthropic>=0.18.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
respx>=0.20.0
